requirements.txt
fastapi==0.115.2
uvicorn==0.32.0
gunicorn==23.0.0
python-dotenv==1.0.0
google-generativeai==0.5.4
psycopg2-binary==2.9.10
